requirements.txt
Flask<1
scikit_learn-0.16.1-cp27-none-win32.whl
scipy-0.16.0-cp27-none-win32.whl
numpy-1.9.2+mkl-cp27-none-win32.whl
matplotlib-1.4.3-cp27-none-win32.whl
